resolut le probleme du update des formation avec les images

+++ server/controllers/formationControllers.js
@@ -22,8 +22,8 @@ const GetFormations = async (req, res) => {
 
 const AddFormation = async (req, res) => {
     const { body } = req
-    if (!body.name || !body.dateDebut || !body.dateFin) {
-        fs.unlinkSync('./public/' + req.file.filename);
+    if (!body.name || !body.dateDebut || !body.dateFin || !req.file) {
+        if (req.file) fs.unlinkSync('./public/' + req.file.filename)
         throw Error('Fill the all fields')
     }
     if (await Formation.findOne({ name: body.name })) throw Error('Formation already exist')
@@ -40,17 +40,20 @@ const AddFormation = async (req, res) => {
 const UpdateFormation = async (req, res) => {
     const { body } = req
     const id = req.params.id
+    let file = ''
     if (!body.name || !body.dateDebut || !body.dateFin) throw Error('Fill the all fields')
     const find_formation = await Formation.findById(id)
     if (!find_formation) throw Error('Formation not existed')
-    const update_formation = await Formation.updateOne({ id }, { ...body })
     if (req.file) {
-        const file = find_formation.image.split('/')[find_formation.image.split('/').length - 1]
-        const update_file_formation = await Formation.updateOne({ id }, { image: `${req.protocol}://${req.get("host")}/${req.file.filename}` })
-        const delete_file = await fs.unlinkSync('./public/' + file)
+        file = find_formation.image.split('/')[find_formation.image.split('/').length - 1]
+        const update_file_formation = await Formation.updateOne({ _id: id }, { ...body, image: `${req.protocol}://${req.get("host")}/${req.file.filename}` })
+        await fs.unlinkSync('./public/' + file)
         if (!update_file_formation) throw Error('Formation not updated try again')
     }
-    if (!update_formation) throw Error('Formation not updated try again')
+    else {
+        const update_formation = await Formation.updateOne({ _id: id }, { ...body })
+        if (!update_formation) throw Error('Formation not updated try again')
+    }
     res.json({ message: 'Successfully, Formation is updated' })
 }
